Allow the value 0 (integer) for our Dropdown
empty(0) returns TRUE even though we would like it to return FALSE
because we consider the integer value of 0 to be a proper value.

The is_int function properly returns FALSE for all non-integer
values (including NULL) so we can use it to catch the case where
our value is 0 and empty should've returned false.